final-pjt/finMunk/finmunk/back/requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.5
atomicwrites==1.4.0
attrs==24.3.0
Automat==25.4.16
babel==2.16.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
black==25.1.0
bleach==6.2.0
blinker==1.9.0
blosc2==3.3.4
boltons==25.0.0
cachetools==6.0.0
catboost==1.2.8
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.3.0
cryptography==43.0.3
cssselect==1.3.0
cycler==0.12.1
datasets==3.6.0
DateTime==5.5
debugpy==1.8.12
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.14
django-allauth==0.61.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.16.0
docstring-to-markdown==0.17
et_xmlfile==2.0.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.2.2
grpcio==1.71.0
gTTS==2.5.4
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.31.1
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
incremental==24.7.2
iniconfig==2.1.0
intel-cmplr-lib-ur==2025.1.1
intel-openmp==2025.1.1
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.5.0
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.9.2
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.6.0
locket==1.0.0
lxml==5.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.3.2
multipledispatch==1.0.0
multiprocess==0.70.16
mypy_extensions==1.1.0
namex==0.0.9
narwhals==1.40.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndindex==1.10.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numexpr==2.10.2
numpy==1.26.3
oauthlib==3.2.2
openai==1.59.7
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.0
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
patsy==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.3.6
plotly==6.1.1
pluggy==1.6.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
Protego==0.4.0
protobuf==4.25.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.13.0
pycosat==0.6.6
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.27.2
pydeck==0.9.1
PyDispatcher==2.0.7
pyerfa==2.0.1.5
pyflakes==3.3.2
Pygments==2.17.2
PyJWT==2.9.0
pykrx==1.0.48
pylint-venv==3.0.4
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyodbc==5.2.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.1.0
python-json-logger==3.2.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.2
python3-openid==3.2.0
pytz==2023.3.post1
PyWavelets==1.8.0
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.0
queuelib==1.8.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.2
requests-file==2.1.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.13.1
Scrapy==2.13.0
seaborn==0.13.2
selenium==4.32.0
Send2Trash==1.8.3
sentencepiece==0.2.0
service-identity==24.2.0
setuptools==80.7.1
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.14.0
tables==3.10.2
tabulate==0.9.0
tbb==2022.1.0
tcmlib==1.3.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
terminado==0.18.1
textdistance==4.6.3
threadpoolctl==3.6.0
tifffile==2025.5.24
tiktoken==0.8.0
tinycss2==1.4.0
tldextract==5.3.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
toolz==1.0.0
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.1
transformers==4.51.3
trio==0.30.0
trio-websocket==0.12.2
Twisted==24.11.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
ultralytics==8.3.119
ultralytics-thop==2.0.14
umf==0.10.0
uri-template==1.3.0
urllib3==2.3.0
w3lib==2.3.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.44.0
widgetsnbextension==4.0.13
Wikipedia-API==0.7.3
wrapt==1.17.2
wsproto==1.2.0
xgboost==2.1.4
xlrd==2.0.1
xxhash==3.5.0
xyzservices==2025.4.0
yapf==0.43.0
yarl==1.18.3
youtube-transcript-api==1.0.3
zipp==3.21.0
zope.interface==7.2
